+++ extend/waibuBootstrap/theme/component/widget/query.js
@@ -63,7 +63,7 @@ async function query () {
       const container = this.params.attr.modal ? 'modal' : 'drawer'
       const scanables = (this.model ? this.model.scanables : []).map(item => req.t(`field.${item}`))
       let placeholder = this.params.attr.placeholder
-      if (!placeholder) placeholder = scanables.length > 0 ? req.t('queryHint%s', join(scanables, { lastSeparator: 'or' })) : req.t('query')
+      if (!placeholder) placeholder = scanables.length > 0 ? req.t('queryHint%s', join(scanables, { separator: ', ', lastSeparator: 'or' })) : req.t('query')
       this.params.html = await this.component.buildSentence(`
         <c:form-input type="search" placeholder="${placeholder}" id="${id}" x-data="{ query: '' }" x-init="
           const url = new URL(window.location.href)